fix: split http instrumentation to core logic and http lib integration
